requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
python-dotenv
itsdangerous
sib-api-v3-sdk
ldap3
bcrypt
google-auth
requests